refactor: Move files to folder 'Systems'
